pyproject.toml
[project]
name = "atlas-board"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.13.0"
dependencies = [
    "django==5.2.4",
    "django-environ==0.12.0",
    "djangorestframework>=3.16.0",
    "mysqlclient==2.2.7",
]
